Add access denied pin
With this you can configure a pin to drive an LED (for example) whenever
the access is denied, either because the card is unkown or because it's
disabled.

It's a way to give a feedback when the lock is not opened. Otherwise you
are not sure if the card wasn't read properly or if your card isn't
enabled.