Adding LED and beeper, plus refactor
In this commit we introduce a LED to signal that we are waiting for the pincode and a beeper to signal if access is denied or not.

Beeper and LED for waiting are going to be used in a later commit.

Together with these new features we are cleaning up a bit the code, in preparation of a refactor of the RFID functionality.