style: standardize whitespace and formatting
The two holy war topics this touches on is spaces vs tabs and if there
should be newlines around braces

omersiar suggested spaces in a github issue, and I chose no no newline
before open brace inside of functions, but not for functions.
I got this from the linux kernel style guide, and I like it because
functions set scope in C, while other blocks do not.

so for example:
void example(void)
{
    if (true) {
        doStuff();
    } else {
        otherStuff();
    }
}