Allow old v1 UID format to work for compatibility
Commits abd0e665 and 20e4dcb7 fixed how UIDs are converted to strings.
Previously some '0's in the UID would be missed, meaning the old stored UIDs
were actually incorrect. This had the effect of meaning esp-rfid v2 is not
compatible with all users from v1. To fix this we create UIDs at scan time in
both the new (fixed) format and the old (broken) format, and check the database
for both.